refactor: rename main file and delete unnecessary file
